requirements.txt
cfgv==3.3.1 ; python_version >= "3.7" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.7" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.7" and python_version < "4.0" and platform_system == "Windows"
distlib==0.3.6 ; python_version >= "3.7" and python_version < "4.0"
filelock==3.8.0 ; python_version >= "3.7" and python_version < "4.0"
flake8-quotes==3.3.1 ; python_version >= "3.7" and python_version < "4.0"
flake8==5.0.4 ; python_version >= "3.7" and python_version < "4.0"
flask==2.2.2 ; python_version >= "3.7" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.7" and python_version < "4.0"
identify==2.5.9 ; python_version >= "3.7" and python_version < "4.0"
importlib-metadata==4.2.0 ; python_version >= "3.7" and python_version < "3.10"
isort==5.10.1 ; python_version >= "3.7" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.7" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.7" and python_version < "4.0"
markupsafe==2.1.1 ; python_version >= "3.7" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.7" and python_version < "4.0"
mypy-extensions==0.4.3 ; python_version >= "3.7" and python_version < "4.0"
mypy==0.991 ; python_version >= "3.7" and python_version < "4.0"
nodeenv==1.7.0 ; python_version >= "3.7" and python_version < "4.0"
platformdirs==2.5.4 ; python_version >= "3.7" and python_version < "4.0"
pre-commit==2.20.0 ; python_version >= "3.7" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.7" and python_version < "4.0"
pycodestyle==2.9.1 ; python_version >= "3.7" and python_version < "4.0"
pyflakes==2.5.0 ; python_version >= "3.7" and python_version < "4.0"
python-dotenv==0.21.0 ; python_version >= "3.7" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.7" and python_version < "4.0"
setuptools==65.6.0 ; python_version >= "3.7" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.7" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.7" and python_version < "3.11"
typed-ast==1.5.4 ; python_version >= "3.7" and python_version < "3.8"
types-psycopg2==2.9.21.2 ; python_version >= "3.7" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.7" and python_version < "4.0"
virtualenv==20.16.2 ; python_version >= "3.7" and python_version < "4.0"
werkzeug==2.2.2 ; python_version >= "3.7" and python_version < "4.0"
yapf==0.32.0 ; python_version >= "3.7" and python_version < "4.0"
zipp==3.10.0 ; python_version >= "3.7" and python_version < "3.10"
boto3